community/finegan/requirements.txt
tensorflow==2.3.0
numpy==1.18.4
absl-py==1.15.0